changed the database file and server file for better readability and used the try and except block
